# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "School Management System API",
    "version": "1.0.0",
    "description": "A comprehensive MERN stack API for managing school operations including authentication, attendance, grades, quizzes, and timetables",
    "contact": {
      "name": "API Support",
      "email": "support@schoolmanagement.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000/api",
      "description": "Development server"
    }
  ],
...
  "paths": {
...
    "/grades/submit": {
      "post": {
        "tags": ["Grades"],
        "summary": "Submit grades",
        "description": "Submit or update grades for a student (Faculty only)",
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["studentId", "classSection", "term", "academicYear", "subjects"],
                "properties": {
                  "studentId": {
                    "type": "string"
                  },
                  "classSection": {
                    "type": "string",
                    "example": "10A"
                  },
                  "term": {
                    "type": "string",
                    "enum": ["Term 1", "Term 2", "Final"],
                    "example": "Term 1"
                  },
                  "academicYear": {
                    "type": "string",
                    "example": "2024-2025"
                  },
                  "subjects": {
                    "type": "object",
                    "additionalProperties": {
                      "type": "object",
                      "properties": {
                        "marks": {
                          "type": "number"
                        },
                        "grade": {
                          "type": "string"
                        }
                      }
                    }
                  },
                  "extraCurricular": {
                    "type": "object"
                  },
                  "descriptiveIndicators": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Grades submitted successfully"
          },
          "403": {
            "description": "Access denied"
          }
        }
      }
    },
    "/grades/report-card/{studentId}": {
      "get": {
        "tags": ["Grades"],
        "summary": "Get report card",
        "description": "Get comprehensive report card with grades and attendance",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "studentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "term",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "academicYear",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Report card retrieved successfully"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  }
}